use PatienChart to create charts
